apps/digidoc/requirements_ocr.txt
# OCR Service Dependencies
# Install with: pip install -r requirements_ocr.txt
#
# Version Pinning Strategy:
# - Production dependencies use == (exact versions) for reproducibility
# - Update versions carefully with testing
# - See development/DEVELOPMENT_BEST_PRACTICES.md for update process

# OCR
pytesseract>=0.3.10  # TODO: Pin to exact version before production

# Image Processing
Pillow>=10.0.0  # TODO: Pin to exact version before production
opencv-python>=4.8.0  # TODO: Pin to exact version before production
numpy>=1.24.0  # TODO: Pin to exact version before production

# Web Framework
flask>=3.0.0  # TODO: Pin to exact version before production
streamlit>=1.28.0  # TODO: Pin to exact version before production

# Utilities
python-dateutil>=2.8.2  # TODO: Pin to exact version before production
pyyaml>=6.0.1  # TODO: Pin to exact version before production
requests>=2.31.0  # TODO: Pin to exact version before production

# Database
sqlalchemy>=2.0.0  # TODO: Pin to exact version before production
alembic>=1.12.0  # Database migrations

# Queue (MVP)
rq>=1.0.0  # TODO: Pin to exact version before production
redis>=5.0.0  # TODO: Pin to exact version before production

# Queue (Post-MVP - commented for now)
# celery>=5.3.0
